Bugfix: 비밀번호 입력 없이 Comment 삭제되는 문제

+++ components/CommentCard.tsx
@@ -224,7 +224,9 @@ const CommentCard = ({ comment, title }: Props) => {
             setIsEditing(false);
           }
         } else {
-          await deleteDoc(commentDocRef);
+          if (password === comment.password) {
+            await deleteDoc(commentDocRef);
+          }
           setIsDeleting(false);
         }
         setPassword("");
